requirements.txt
tensorflow
numpy
matplotlib
Pillow
scikit-learn
flask
jupyter